+++ test/suite/extension.test.ts
@@ -10,16 +10,24 @@ const fixtureDir = path.resolve(
 
 suite('Extension Tests', () => {
   test('cargo tasks are auto-detected', async () => {
-    const projectPath = path.join(fixtureDir, 'bare-lib-project');
+    const projectPath = fixtureDir;
     const projectUri = Uri.file(projectPath);
+    const projects = [
+      path.join(projectPath, 'bare-lib-project'),
+      path.join(projectPath, 'another-lib-project'),
+    ];
 
     await vscode.commands.executeCommand('vscode.openFolder', projectUri);
     await vscode.workspace.openTextDocument(
-      Uri.file(path.join(projectPath, 'src', 'lib.rs')),
+      Uri.file(path.join(projects[0], 'src', 'lib.rs')),
+    );
+    await vscode.workspace.openTextDocument(
+      Uri.file(path.join(projects[1], 'src', 'lib.rs')),
     );
 
     const expected = [
-      { subcommand: 'build', group: vscode.TaskGroup.Build },
+      { subcommand: 'build', group: vscode.TaskGroup.Build, cwd: projects[0] },
+      { subcommand: 'build', group: vscode.TaskGroup.Build, cwd: projects[1] },
       { subcommand: 'check', group: vscode.TaskGroup.Build },
       { subcommand: 'test', group: vscode.TaskGroup.Test },
       { subcommand: 'clean', group: vscode.TaskGroup.Clean },
@@ -28,13 +36,18 @@ suite('Extension Tests', () => {
 
     const tasks = await vscode.tasks.fetchTasks();
 
-    for (const { subcommand, group } of expected) {
+    for (const { subcommand, group, cwd } of expected) {
       assert(
         tasks.some(
           task =>
             task.definition.type === 'cargo' &&
             task.definition.subcommand === subcommand &&
-            task.group === group,
+            task.group === group &&
+            (!cwd ||
+              cwd ===
+                (task.execution &&
+                  task.execution.options &&
+                  task.execution.options.cwd)),
         ),
       );
     }